# docs/openapi/raven-services.yaml
openapi: "3.1.0"
info:
  title: Raven AI — Microservice API
  version: "1.0.0"
  description: |
    Enterprise-grade personal AI assistant.
    Monolith API at :18888, microservices behind gateway at :8000.
servers:
  - url: http://localhost:18888/api
    description: Monolith API
  - url: http://localhost:8000/api/v1
    description: Gateway (microservice mode)
paths:
...
  /auth/register:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string, minLength: 3 }
                password: { type: string, minLength: 8 }
      responses:
        "201": { description: User created }
        "400": { description: Invalid input }
        "409": { description: Username exists }
  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expires_at: { type: integer }
                  user_id: { type: string }
  /auth/validate:
    post:
      summary: Validate JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  user_id: { type: string }
                  role: { type: string }

  # ── Agent Core ────────────────────────────────────────
...
  /code/execute:
    post:
      summary: Execute code in sandbox
      tags: [Code]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                language: { type: string }
                code: { type: string }
      responses:
        "200": { description: Execution result }

  # ── Admin ─────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...